Refactored and restructured the code
The code was refactored and restructured in order to follow the directives of using React.

All components are now functional components  and can now be called upon to be re-used. Applied context  hooks and rules in order to keep global states and pass necessary variables to the separate functional components.